+++ gtk/buddylookup.c
@@ -178,8 +178,8 @@ static gboolean linphone_gtk_process_buddy_lookup(GtkWidget *w){
 					results);
 			gtk_progress_bar_set_fraction(pb,1);
 			tmp=g_strdup_printf(ngettext("Found %i contact",
-                        "Found %i contacts", ms_list_size(results)),
-                    ms_list_size(results));
+                        "Found %u contacts", ms_list_size(results)),
+                    (unsigned int)ms_list_size(results));
 			gtk_progress_bar_set_text(pb,tmp);
 			g_free(tmp);
 			sip_setup_context_buddy_lookup_free(ctx,req);
